refactor basic style
