+++ src/App.js
@@ -1,13 +1,13 @@
 import React from 'react'
 import { GlobalStyle } from './styles/GlobalStyle'
 import { ProductsContainer } from './containers/ProductsContainer'
-import { Title } from './styles/Layout'
+import { Title } from './styles/GlobalComponents'
 
 export const App = () => {
   return (
     <div>
       <GlobalStyle />
-      <Title>Product List</Title>
+      <Title>Produktlista</Title>
       <ProductsContainer />
     </div>
   )
